Move TopicHelpPage to interfaces layer to fix import-linter violation
The register_help_topics hookspec (plugins layer) imported TopicHelpPage from
cli.help_topics, but the import-linter layers contract places cli ABOVE plugins,
so plugins must not import from cli. Move the plugin-facing surface
(TopicHelpPage model + build_topics_from_directory helper) to a new
interfaces/help_topic.py -- the interfaces layer is below plugins and already
holds plugin-facing models like CreateAgentOptions. The runtime topic registry
(get_topic/get_all_topics/register_topic/preserve_topic_registry) stays in
cli/help_topics.py. Registration mutation moved out of TopicHelpPage.register()
into register_topic so the model no longer references the global registry.
