+++ libs/mngr/changelog/mngr-help-docs-plugins.md
@@ -1,3 +1,3 @@
 Plugins can now contribute standalone help topic pages via the new `register_help_topics` hook. Topics from an installed plugin show up in `mngr help` and are viewable via `mngr help <topic>`, just like mngr's built-in topics. Implement the hook to return `TopicHelpPage` objects (for aliases, "See Also" references, and inline content), or use the new `build_topics_from_directory` helper to expose a directory of markdown files as topics (one per file). Plugin topics that collide with a built-in topic key are skipped so built-in topics always win.
 
-Internally, the topic data model (`TopicHelpPage`), the topic registry (`get_topic`/`get_all_topics`), and the new `build_topics_from_directory` helper now live in a lightweight `imbue.mngr.cli.help_topics` module so plugins can import them without pulling in the rest of the CLI. mngr's own built-in topics are now registered through this same hook (as a built-in plugin), the same way built-in provider backends and agent types register through their hooks.
+Internally, the plugin-facing `TopicHelpPage` model and the `build_topics_from_directory` helper live in `imbue.mngr.interfaces.help_topic` (so the plugin hookspec can reference them without the plugins layer importing the CLI), while the runtime topic registry lives in `imbue.mngr.cli.help_topics`. mngr's own built-in topics are now registered through this same hook (as a built-in plugin), the same way built-in provider backends and agent types register through their hooks.
